+++ seantis/people/supermodel/__init__.py
@@ -13,4 +13,4 @@ from .indexing import (
     get_selectable_field_ix
 )
 
-from .schema_columns import get_schema_columns
+from .schema_columns import get_schema_columns, unrestricted_get_schema_columns