+++ seantis/people/tests/test_schema_columns.py
@@ -1,7 +1,9 @@
 from plone.supermodel import loadString
 
 from seantis.people import tests
+from seantis.people.supermodel import security
 from seantis.people.supermodel import (
+    unrestricted_get_schema_columns,
     get_schema_columns,
     set_columns,
     set_title_fields
@@ -32,7 +34,7 @@ class TestSchemaColumns(tests.IntegrationTestCase):
     def test_no_columns(self):
 
         model = loadString(self.address_model)
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
 
         self.assertEqual(columns, [])
 
@@ -41,7 +43,7 @@ class TestSchemaColumns(tests.IntegrationTestCase):
         model = loadString(self.address_model)
 
         set_columns(model.schema, [['lastname']])
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
 
         self.assertEqual(len(columns), 1)
         self.assertEqual(columns[0].schema, model.schema)
@@ -54,7 +56,7 @@ class TestSchemaColumns(tests.IntegrationTestCase):
         model = loadString(self.address_model)
 
         set_columns(model.schema, [['country', 'lastname']])
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
 
         self.assertEqual(len(columns), 1)
 
@@ -69,17 +71,44 @@ class TestSchemaColumns(tests.IntegrationTestCase):
 
         set_columns(model.schema, [['country', 'lastname']])
 
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
         self.assertEqual(columns[0].contains_title_field, False)
 
         set_title_fields(model.schema, ['lastname'])
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
         self.assertEqual(columns[0].contains_title_field, True)
 
         set_title_fields(model.schema, ['firstname'])
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
         self.assertEqual(columns[0].contains_title_field, False)
 
         set_columns(model.schema, [['firstname']])
-        columns = get_schema_columns(model.schema)
+        columns = unrestricted_get_schema_columns(model.schema)
         self.assertEqual(columns[0].contains_title_field, True)
+
+    def test_restricted_columns(self):
+
+        context = self.new_temporary_folder()
+        model = loadString(self.address_model)
+
+        set_columns(model.schema, [
+            ['firstname', 'lastname'],
+            ['town'],
+            ['country'],
+        ])
+        security.set_read_permissions(model.schema, {
+            'firstname': 'cmf.ManagePortal',
+            'lastname': 'zope2.View'
+        })
+
+        # anonymous
+        columns = get_schema_columns(model.schema, context)
+        self.assertEqual(len(columns), 3)
+        self.assertEqual(columns[0].fields, ['lastname'])
+
+        # admin
+        with self.user('admin'):
+            columns = get_schema_columns(model.schema, context)
+
+        self.assertEqual(len(columns), 3)
+        self.assertEqual(columns[0].fields, ['lastname', 'firstname'])