+++ docs/HISTORY.rst
@@ -2,6 +2,12 @@
 Changelog
 ---------
 
+0.5 (unreleased)
+~~~~~~~~~~~~~~~~
+
+- Support supermodel security permissions in the people's list (giving the
+  ability to hide certain fields in the table depending on the user)
+
 0.4
 ~~~